+++ CHANGELOG.md
@@ -8,6 +8,13 @@ Run Configurations:
     query processor.
 1.  Support profiling Saxon XSLT and XQuery scripts.
 
+Saxon:
+
+1.  Fix using Saxon 9.2 to 9.8 JAR files: the `RawDestination` class is not
+    available in these versions.
+1.  Disable Saxon EE optimizations to prevent the processor throwing a
+    `NoClassDefFoundError` looking for `com/saxonica/ee/bytecode/GeneratedCode`.
+
 ## 1.4.1 - 2019-03-27
 
 IntelliJ Integration: